pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "academy-learning-agent"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
repository = "https://github.com/valory-xyz/academy-learning-agent"
keywords = [ "autonomy", "autonomous", "autonolas", "service", "multi", "agent",]
classifiers = [ "Environment :: Console", "Environment :: Web Environment", "Development Status :: 2 - Pre-Alpha", "Intended Audience :: Developers", "License :: OSI Approved :: Apache Software License", "Natural Language :: English", "Operating System :: MacOS", "Operating System :: Microsoft", "Operating System :: Unix", "Programming Language :: Python :: 3.8", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Topic :: Communications", "Topic :: Internet", "Topic :: Scientific/Engineering", "Topic :: Software Development", "Topic :: System",]
[[tool.poetry.packages]]
include = "packages"

[tool.poetry.dependencies]
python = "<4.0,>=3.10"
open-autonomy = "==0.16.1"
open-aea = "==1.57.0"
open-aea-ledger-ethereum = "==1.57.0"
typing_extensions = ">=3.10.0.2"
toml = "==0.10.2"
tomte = {version = "==0.2.17", extras = ["cli", "tests"]}
black = "==24.2.0"
isort = "==5.13.2"
grpcio = "==1.53.0"
asn1crypto = "<1.5.0,>=1.4.0"
open-aea-ledger-cosmos = "==1.57.0"
py-ecc = "==6.0.0"
pytz = "==2022.2.1"
openapi-core = "==0.15.0"
openapi-spec-validator = "<0.5.0,>=0.4.0"
click = "^8.1.7"
